Improve behavior of Homebrew package manager when installed somewhere other than /usr/local, which is the default location for installing Homebrew.
This includes a tweak with the temporary directory for Homebrew, which
needs to be in the same volume as the Homebrew install location for
certain packages like git.

It also includes another tweak to preserve an externally linked vimrc
configuration file, which tends to get blown away by simply installing
upgrades of the vim package, or reinstalling it.